Removed table_catalog check
...from macro list_relations_without_caching.
The problem is that the dbname that dbt uses to connect ({{ schema_relation.database }}), is not necessarily the same as SYS_CONTEXT('userenv', 'DB_NAME'), and since the table_catalog column is generated anyway, the result is that one gets an all-or-nothing list of tables in the right schema, depending on whether dbname matches the database name. A symptom of this problem is that dbt can't find tables to drop before it tries to recreate them.